Backend/AITrip/requirements.txt
Django==5.2

djangorestframework==3.16.0

django_cors_headers==4.7.0

django_filter==25.1

djangorestframework_simplejwt==5.5.0

mysqlclient==2.2.4

django-filter==25.1
